refactor: PraiseSection としてコンポーネント化
